working new instance list
