+++ hat/assets/js/apps/Iaso/routing/redirections.js
@@ -3,8 +3,6 @@ import { Redirect, Route } from 'react-router';
 import { baseUrls } from '../constants/urls';
 import { getChipColors } from '../constants/chipColors';
 
-import { getMetasColumns } from '../domains/instances/utils';
-
 import { locationLimitMax } from '../domains/orgUnits/constants/orgUnitConstants';
 import Page404 from '../components/errors/Page404';
 
@@ -18,10 +16,6 @@ const addRoutes = baseRoutes =>
             }/locationLimit/${locationLimitMax}/order/id/pageSize/50/page/1/searchTabIndex/0/searches/[{"validation_status":"all", "color":"${getChipColors(
                 0,
             ).replace('#', '')}"}]`}
-        />,
-        <Redirect
-            path={baseUrls.instances}
-            to={`${baseUrls.instances}/columns/${getMetasColumns().join(',')}`}
         />,
         <Redirect
             path={baseUrls.mappings}
